+++ utils/commands.ts
@@ -145,6 +145,13 @@ const COMMANDS: Command[] = [
     hidden: true,
     isEnabled: ALWAYS_ENABLED,
   },
+  {
+    id: 'toggle-table-colors',
+    label: 'Toggle Table Colors',
+    category: 'Preferences',
+    handler: global => global.togglePineTableColors(),
+    isEnabled: ALWAYS_ENABLED,
+  },
 
   // Experimental Category
   {
@@ -163,6 +170,7 @@ const COMMANDS: Command[] = [
     handler: global => global.setShowAnalysis(true),
     isEnabled: ALWAYS_ENABLED,
   },
+
 
   // Help Category
   {